Spring Data JPA [#505127]

+++ src/main/resources/db/scripts/002_ddl_create_posts_table.sql
@@ -3,11 +3,11 @@
 --changeset devvk:002_ddl_create_posts_table
 CREATE TABLE posts
 (
-    id         SERIAL PRIMARY KEY,
+    id         BIGSERIAL PRIMARY KEY,
     title      VARCHAR(255) NOT NULL,
     body       TEXT         NOT NULL,
     created_at TIMESTAMP    NOT NULL DEFAULT CURRENT_TIMESTAMP,
-    user_id    INT          NOT NULL REFERENCES users (id)
+    user_id    BIGINT       NOT NULL REFERENCES users (id)
 );
 
 --rollback DROP TABLE posts;
